requirements.txt
django
djangorestframework
django-filter
drf-yasg
python-dotenv
requests
gunicorn
django-cors-headers
pytest
pytest-django
pytest-cov
drf-nested-routers
